feat: organization access control

+++ lib/dtt_recharger_web/controllers/organization_controller.ex
@@ -1,29 +1,39 @@
 defmodule DttRechargerWeb.OrganizationController do
   use DttRechargerWeb, :controller
+  alias DttRecharger.Policies.OrganizationPolicy
 
   alias DttRecharger.Operations.OrganizationOperation
   alias DttRecharger.Schema.Organization
 
   def index(conn, _params) do
-    organizations = OrganizationOperation.list_organizations()
-    render(conn, :index, organizations: organizations)
+    case  OrganizationPolicy.index(conn.assigns.current_user_role) do
+      true -> render(conn, :index, organizations: OrganizationOperation.list_organizations())
+      false -> redirect(conn, to: ~p"/")
+    end
   end
 
   def new(conn, _params) do
-    changeset = OrganizationOperation.change_organization(%Organization{})
-    render(conn, :new, changeset: changeset)
+    case  OrganizationPolicy.index(conn.assigns.current_user_role) do
+      true -> render(conn, :new, changeset: OrganizationOperation.change_organization(%Organization{}))
+      false -> redirect(conn, to: ~p"/")
+    end
   end
 
   def create(conn, %{"organization" => organization_params}) do
-    case OrganizationOperation.create_organization(organization_params) do
-      {:ok, organization} ->
-        conn
-        |> put_flash(:info, "Organization created successfully.")
-        |> redirect(to: ~p"/organizations/#{organization}")
+    if OrganizationPolicy.index(conn.assigns.current_user_role) do
+      case OrganizationOperation.create_organization(organization_params) do
+        {:ok, organization} ->
+          conn
+          |> put_flash(:info, "Organization created successfully.")
+          |> redirect(to: ~p"/organizations/#{organization}")
 
-      {:error, %Ecto.Changeset{} = changeset} ->
-        render(conn, :new, changeset: changeset)
+        {:error, %Ecto.Changeset{} = changeset} ->
+          render(conn, :new, changeset: changeset)
+      end
+    else
+      redirect(conn, to: ~p"/")
     end
+
   end
 
   def show(conn, %{"id" => id}) do
@@ -32,15 +42,23 @@ defmodule DttRechargerWeb.OrganizationController do
   end
 
   def my_organization(conn, _params) do
-    org = conn.assigns.current_organization
-    organization = OrganizationOperation.get_organization!(org.id)
-    render(conn, :my_organization, organization: organization)
+    if OrganizationPolicy.index(conn.assigns.current_user_role) do
+      org = conn.assigns.current_organization
+      organization = OrganizationOperation.get_organization!(org.id)
+      render(conn, :my_organization, organization: organization)
+    else
+      redirect(conn, to: ~p"/")
+    end
   end
 
   def edit(conn, %{"id" => id}) do
-    organization = OrganizationOperation.get_organization!(id)
-    changeset = OrganizationOperation.change_organization(organization)
-    render(conn, :edit, organization: organization, changeset: changeset)
+    if OrganizationPolicy.index(conn.assigns.current_user_role) do
+      organization = OrganizationOperation.get_organization!(id)
+      changeset = OrganizationOperation.change_organization(organization)
+      render(conn, :edit, organization: organization, changeset: changeset)
+    else
+      redirect(conn, to: ~p"/")
+    end
   end
 
   def update(conn, %{"id" => id, "organization" => organization_params}) do
@@ -58,11 +76,15 @@ defmodule DttRechargerWeb.OrganizationController do
   end
 
   def delete(conn, %{"id" => id}) do
-    organization = OrganizationOperation.get_organization!(id)
-    {:ok, _organization} = OrganizationOperation.delete_organization(organization)
+    if OrganizationPolicy.index(conn.assigns.current_user_role) do
+      organization = OrganizationOperation.get_organization!(id)
+      {:ok, _organization} = OrganizationOperation.delete_organization(organization)
 
-    conn
-    |> put_flash(:info, "Organization deleted successfully.")
-    |> redirect(to: ~p"/organizations")
+      conn
+      |> put_flash(:info, "Organization deleted successfully.")
+      |> redirect(to: ~p"/organizations")
+    else
+      redirect(conn, to: ~p"/")
+    end
   end
 end
